Adding variadic tempalte function to class ADIOS
Adding variadic functions Open and Write to pass different options as
different transport will require different arguments.

Rewriting class CCapsule, now representing a stream with a buffer,
metadata and a collection of CTransports.

Write API would focus on Streams, not Group. This is for consistency
with a Read API.

To do:

Finish reorganizing the code for compilation.