Place fixed width fundamental types into the adios namespace
This allows us to both drop the namespace prefix when using them in
implementation and guarantee that we don't use versions of the types
that get re-defined elsewhere.

In order to provide the least ammount of disruption this doesn't include
any sweeping search and replace changes.  However, it is recommended
moving forward that these types be used internally when fixed width
sizes are expected.  This will allow the transition to happen gradually
through attrition and minimize conflicts.