Add type traits to map C type info to fixed width types.
The type traits will be useful for maping ambiguous C types to fixed
width integer types used for the actual I/O operations, for instance:

  adios::TypeInfo<char>::IOType resolves to int8_t
  adios::TypeInfo<signed char>::IOType resolves to int8_t
  adios::TypeInfo<unsigned signed char>::IOType resolves to uint8_t
  adios::TypeInfo<long int>::IOType resolves to int64_t
  adios::TypeInfo<long long int>::IOType resolves to int64_t

So in this case, even though char and signed char are "the same", they
are distinctly separate types to the compiler.  Using
adios::TyperInfo<T>::IOType you can use the same fundamental type for all
of them.  Similarly for long int and long long int.