Fix Fortran module race condition in build of tests
Place SmallTestData_mod.f90 into an object library to avoid multiple copies of
small_test_data.mod from overwriting one another in parallel builds.  CMake
should be able to handle this on it's own a bit more inteligently but at the
moment it does not so the object library approach used here is really just a
workaround but should be sufficient for what we need.