Remove try-catch from main in tests
The initial intent of the try{}catch(exception){} in the test code mains was
to squelch a large number of issues flagged by the Coverity static analyzer.
Unintentionally, this introduced a new set of warnings on the Windows
builds and still didn't address the Coverity warnings.  Since the uncaught
exception is intended behavior by the GTest framework, this restores the
original exception throwing functionality so the issue can be silenced in
Coverity through an appropriate modeling file.