cxx11-bindings: mv adios2::GetType<T>() into cxx11 interface only
For the most part, the cxx11 interface is well separate from the core code
by means of core code being in the adios2::core namespace. However, some
code is shared, between core and the cxx11 interface, in particular, the
type definitions like ShapeID. That makes sense, as one probably doesn't
anticipate the use of those types do diverge between the cxx11 bindings and
the core.

Howevere, the GetType function was apparently kinda designed to be
potentially different between core and cxx11 interface, because there are
two versions of it. And that's actually helpful now. So core code should
use helper::GetType (which maybe should be called core::GetType) only.
That was mostly the case, but the top-level adios2::GetType was also
available to core as it was in the shared ADIOSTypes.h and there were
a few instances that used it, which I fixed in the previous commit.

I now moved adios2::GetType it into cxx11/, so it can't be accidentally used
by the core anymore.