cxx11-bindings: adios2::GetType<T>() maps to IOType first
If we create a Variable<unsigned long>, we'll actually get one that
wraps the core::Variable<uint64_t> (assuming unsigned long is 64 bit,
otherwise it'd be uint32_t). So GetType<T>() is accordingly be redefined
so that GetType<unsigned long>() gives "uint64_t".

Well, eventually -- right now the string used is actually
"unsigned long long int" for compatibility with how the types have been
named in the past.