stdtype: engine/bp3
Convert engine/bp3 to only handle to stdint-based types that I'm
using the moniker "STDTYPE" for.

Eventually, everything will be on "STDTYPE" only, so it can then be
renamed back to TYPE.

Engines interface directly with the CXX11/C bindings, which are already
using only the STDTYPEs. On the other side, they interface with e.g.,
formats, which in practice only deal with Variables of STDTYPEs only, too.

If they didn't, changes like this would lead to unresolved references
(unistantiated templates), so if things were to break, it'd be obvious.