EXP: change API for adios2_variable_name
This does break the existing API.

I adapted three uses, the test for it,
the Fortran bindings and the Matlab bindings (the latter without even
compiling it). Obviously, there might be external users out there.
The test and the Matlab bindings show that the new interface
is much more natural to use from C.