bpls: fix uninitialized arg parsing bit
It looks to me like `process_unused_args()` adds a dummy arg0 into
argv which is created by `new char[4]`. This gets later processed
and crashes when `strlen()` is called on it, since it's not necessarily
0-terminated. (Because it's not initialized at all.) This is bug is
triggered by clang's address sanitizer, and fix by initializing the
allocated memory to 0 here.