Fixes #1253 (master broken with Endian_Reverse)
CopyPayloadStride was using ADIOS2_FOREACH_PRIMITIVE_TYPE_1ARG, which
includes non-stdint types like char, for which GetType<> isn't
specialized anymore.

Fixing that to use ADIOS2_FOREACH_PRIMITIVE_STDTYPE then leads to another
issue, that is missing explicit instantiation of CopyEndianReverse,
which is defined in adiosMemory.tcc, and therefore needs to be instantiated
explicitly.

CopyPayloadStride uses CopyEndianReverse for (now) the stdint-based types,
so that takes care of most of what's needed, but `char` is missing,
so this patch explicitly instantiates the CopyEndianReverse<char>.