rename adios2_engine_type -> adios2_io_engine_type
Its Fortran version is already named that way, and it makes more sense,
since the function actually takes an adios2_io, not an adios2_engine.
This will allow to later add a adios2_engine_type function to the C
bindings that actually takes an adios2_engine (if needed). (The CXX11
interface already has Engine::Type(), which does not have an equivalent
 in the C or Fortran bindings).