mv the ADIOS2_FOREACH_*_TYPE_1ARG macros -> bindings/CXX11
These macros use primitive-based integers, which are only ever to be
used in the bindings/CXX11 interface, while the rest of ADIOS2 should
(and already does) use the _STDTYPE_ version. In order to avoid the
_TYPE macros being mistakenly used in the core, this patch moves them
into bindings/CXX11/cxx11/Types.h, which gets installed. The macros
need to be part of the installed interface, as they were previously
through ADIOSMacros.h, as they are used in header files.