work around codacy issue
It's complaining that "Variable 'attributesDataMap' is assigned a value
that is never used.", which isn't actually true. Maybe it gets confused
by the macros, though that doesn't make much sense to me, as I'd think
it analyzes code after running it through the preprocessor.

Anyway, hopefully this will work around it.