Fix internal state corruption in aggregator
The absolute position exchange is currently implemented in such a way that only allows a single operation to be in flight at any give time.  This unnecessarily restricts the base implementation so this moves it into the child class and enforces in that implementation that it's only called one at a time.