mpiwrap: provide SMPI_* wrappers around MPI that handle serial case
In the serial case, these wrappers will always call through to the mpidummy
MPI functons.

In the case where adios2 is compiled with MPI support, the SMPI wrappers will
call through to the actual MPI implementation, if MPI has been initialized
(which is currently always the case, otherwise one would get errors from the
MPI library). The new behavior is that the wrappers catch the case where MPI
has not been initialized and fall back to mpidummy in this case.