mpiwrap: use MPI_COMM_NULL to indicate "serial" ADIOS2
Switching between real MPI and mpidummy based on whether MPI was
initialized was not really clean. That decision should be made based
upon whether the user is using the serial API (not passing any
communicator) or the parallel API (when passing a communicator).

Also, it's possible and not even entirely unlikely to start using ADIOS2
in serial mode and then after MPI_Init got called later, it'd switch into
parallel mode, which actually really happened, that is, if there is a
statically declared ADIOS object, that will be initialized before main(),
so before MPI_Init. That's why in ADIOS.cpp I didn't use
SMPI_Comm_dup/free, because that lead to erros.

This patch enables ADIOS2 to switching behavior based upon whether the
serial or parallel API is invoked, by internally using the special value
`MPI_COMM_NULL` to remember which mode we're operating in.