Skip first argument when generating config file
Linking against ADIOS from a C program requires linking with a C++
linker. However, this causes problems with generating the config
file, because the actual compiler used shows as a difference between
the flags, resulting in the C++ compiler being one of the "flags" that
is erroneously emitted by the config file. To remedy this, skip the
first argument when doing the comparison between the two sets of flags.