Use CTest fixtures in Install tests
Fixtures were initially not used because Adios supports CMake 3.6,
which did not yet have fixtures (they were introduced in 3.7.)
However, adding the fixtures can be done in a backwards compatible
manner without affecting pre-3.7 builds. Add test fixtures to ensure
that Install.Setup always gets run before the other Install tests, even
if it's not specified in the CTest -R regex. (Note that we are still
using the DEPENDS property in addition to FIXTURES_REQUIRED in order
to support 3.6.)