python/hl: fix read() with selection of LocalArrays
This fixes the two previously expectedFailure test cases, though only after
fixing the test cases (they previously broke not because of the incorrect
reference data, but because they threw exceptions).

Also fixes the error message that confused me into thinking the problem was
with core, and a couple of similar cases.