layout: add support for col-major (in addition to row-major)
An additional, optional, `order` argument to read will determine how
the data that is being read is represented as a numpy array. Options
are 'C', 'F', 'K', which are row-major, column-major and original layout,
respectively. If original layout is not known, 'C' will be used, which
is previous behavior.

If the order argument is not provided, the read will default to
`adios2.File.read_order`, which itself defaults to 'K', but user code
can override that default.