core: Register MPI engines separately
Split registration of IO factory entries.  Register serial engines
statically.  Add runtime registration of MPI engines that are
guaranteed to be available if the application creates a MPI-backed
ADIOS communicator.

The HDF5 engine is special in that it is always available to programs
using serial communicators but is available to those using MPI
communicators only if HDF5 was built with parallel support.  Teach the
HDF5 engine factory function to create the engine only when using an
appropriate communicator.

The SST engine also supports working with or without MPI.  Leave it
registered statically for now.  Later the two variants can be split.