table: Enable engine without MPI
Now that we manage table's sub-adios through the core API we can access
the core ADIOS constructor that accepts the Comm abstraction instead of
a raw MPI_Comm.  We can use `Comm::World()` instead of `MPI_COMM_WORLD`.
This allows us to enable the Table engine even when ADIOS is built
without MPI.