helper: Update AvailableIpAddresses to satisfy sanitizers
Fix end condition when iterating through `if_nameindex` array.
Add missing resource free operations in error conditions.

Unfortunately the Clang Memory Sanitizer does not recognize that
`if_nameindex` initializes the array behind the pointer it returns
and complains on the first dereference that it is uninitialized.
Suppress the Memory Sanitizer for this function.