helper: Always provide adiosNetwork functions on UNIX platforms
Since commit 19545e6d9 (staging part worked in table engine, 2019-04-13,
v2.5.0~80^2~14) the `AvailableIpAddresses` function is needed by the
Table engine but its conditions only make it available for dataman or
ssc.  Drop those conditions so it is available in general.

For now we still need to leave these out on Windows.  The
`AvailableIpAddresses` function in particular needs a conditional
alternative implementation on Windows.