bindings: Fix Python serial program init when ADIOS2 has MPI enabled
The one-argument `adios2.ADIOS(...)` constructor has two overloads
when MPI is enabled: `ADIOS(bool)` and `ADIOS(MPI4PY_Comm)`.  The
representation of `MPI4PY_Comm` is just an integer, so a call with
a `bool` may try to convert it (and fail).  Re-order the pybind11
definitions of these signatures so that the `ADIOS(bool)` is tried
first.  That way a call with a `bool` will use it and a call with
a `MPI4PY_Comm` will not convert and fall back to the proper variant.

Fixes: #2233