Enforce invariants for inline engine
The inline engine must have exactly one writer, and exactly one reader.
If a user attempts to add multiple readers or multiple writers, throw an exception.
This is a first step towards unifying the engine APIs, since currently we require
the user to specify which engine is the reader with a key.

Enforcing this invariant will allow us to remove the parameter keys, since we will
be able to uniquely identify the reader and writer via their respective modes.

As a bit of unrelated cleanup: throw exceptions when users request unsupported features,
like append mode or sync mode for the inline engine.