bindings: Help C++11 clients avoid calling serial constructor with MPI_Comm
Client code that does not enable `ADIOS2_USE_MPI` may accidentally try to pass
a `MPI_Comm` instance to our constructor.  If the type the MPI library uses to
implement MPI_Comm happens to be convertible to one of the types offered by our
serial constructors (e.g. `bool`), the invocation will appear to succeed but
will ignore the communicator passed.  Add explicitly deleted constructors that
have better conversions for common `MPI_Comm` types.

With this, an attempt in client code to call the `adios2::ADIOS` constructor
with a `MPI_Comm` object will fail to compile with an error such as:

    error: use of deleted function 'adios2::ADIOS::ADIOS(void*)'
        adios2::ADIOS adios(comm);
    ...note: declared here
        ADIOS(void *define_ADIOS2_USE_MPI_to_use_MPI_constructor) = delete;
        ^~~~~