FindBlosc.cmake: Threads
Blosc depends unconditionally on threads.
If we are not using the config package, we need to re-add the
pthreads flags or we will get linker errors.

We are sometimes lucky at the moment that in some build configs
pthread flags land in the right order on the commands line, but
generally we should declare this dependency.