ZFP: Fixed odd test params: accuracy->rate
This was an oddity in the ZFP unit tests since the name of the tests and
its suite are `rate` something, but some of them hiddenly used accuracy
instead of rate.

Moreover, the value given to accuracy was the same value as the rate
which is always > 1, while accuracy must be by definition less the 1,
this that perhaps worked in previous versions of ZFP correcting these
input values but it is definitely not that case for current versions.
This commit resolves that and their associate errors.

Also some refactor at the compressZFP implementation.