SST: Add UCX SST Dataplane
This commit implements a UCX SST dataplane.

The UCX dataplane was added without support for PRELOAD functionality as discussed previously.

Another open topic is the priority of the UCX dataplane (compared to other DPs), if others are available. Currently we are hardcoding the priority to a value of 10.

The dataplane was tested with UCX v1.11 and openmpi v4.1.4 (compiled with UCX support) with the following cases:
  1. SST ctests.
      99% tests passed, 1 tests failed out of 168
          The following tests FAILED:
          1056 - Engine.Staging.TestOnDemandMPI.ADIOS2OnDemandMPI.Sst.MPI (Timeout)
  2. Gray Scott from ADIOS2-EXAMPLES

Signed-off-by: Sameeh Jubran <sameeh.j@gmail.com>
Signed-off-by: Michael Laufer <mlaufer103@gmail.com>
Signed-off-by: Erick Fredj <efredj@toganetworks.com>
Signed-off-by: Mark Wasserman <mark.wasserman@toganetworks.com>