ReadModeFlatten (#4147)
Add a new external mode for opening files, ReadFlattenSteps.  This is like ReadRandomAccess (in that it reads all metadata immediately and you can't use BeginStep), but it causes *all* the timesteps in the file to be Flattened into a single step, as if each Put() has been done in that step.