Introduce DataSizeBased aggregation method
Add a new aggregation method which behaves just like the existing
EveryoneWritesSerial method, excepts builds the writer chains after
first partitioning ranks in such a way that each chain writes
approximately the same amount of data.

The current partitioning method is a greedy approach that is fast
but not garanteed to produce optimal partitionings. However, this
has been done in such a way that adding new partitioning methods
should be straightforward.