Fix issue with GCC 7.
This version of GCC came out around the same time as C++17. Although
its preprocessor macros claim that it supports the C++17, there were
still some portions of the specifications that were not supported.
In particular, `std::filesystem` is not supported. In there case where
this is needed, detect GCC 7 and do not attempt to use it.