Fix race condition
Cause of race condition was that rank 0 would, maybe due to its extra
pre-write duties, arrive to the party and send its WRITE_SUBMISSION
msg after the SC had already finished writing and exited the comm
loop.

Since all ranks generate the complete partitioning information, we
can make that available so that SCs can pre-populate their writer
queues from that and do away with the WRITE_SUBMISSION message
altogether. This reduces the total number of messages needed, as
well as avoids the race condtion that caused intermittent hangs
of the test.