Fix failing tests when using the new rerouting feature
* Before leaving the comm loop, SC must update its m_DataPos from the
variable that has been tracking it for its group, or else it shares
the wrong file positions with the other ranks. This was causing all
tests with more than 1 timestep to fail.

* When SC comm thread first starts up, it needs to set the current
file position variable from m_DataPos, which was updated correctly
in append mode.  That was causing tests that append to fail.