Fix subtle bug that could cause blocks to disappear
This could happen when global coordinator re-routed another rank
to it's own subfile. Because the update of m_DataPos from the local
tracking varibale, currentFilePos, was happening upon receipt of
the GROUP_CLOSE message, and the global coordinator does not send
itself that message, it wasn't capturing the actual file offset to
be shared with other ranks later in EndStep().

This moves the variable update to the end of the thread method,
just before returning, so that all subcoordinators do it, regardless
of whether they're the global coordinator also.